#51 Refactor: Team 앱 Custom Manager 작성 스크립트 추가
